update console.log to more meaningful text

+++ src/components/views/game/MultiPlayerGame/CreatedGamePage.js
@@ -32,13 +32,9 @@ const CreatedGamePage = () => {
   const history = useHistory();
 
   const fetchPlayer = async () => {
-    try {
-      const myValue = localStorage.getItem("isServer");
-      console.log(isServer);
-      const response = await api.get(
-        `/games/${localStorage.getItem("gameId")}/players`
-      );
-      console.log("Players", response.data);
+    try{
+      const response = await api.get(`/games/${localStorage.getItem("gameId")}/players`);
+      console.log("Players List: ", response.data);
       setGamePlayers(response.data);
       setPlayerNumber(response.data.length);
     } catch (error) {
@@ -53,16 +49,23 @@ const CreatedGamePage = () => {
       const response = await api.get(
         `/games/${localStorage.getItem("gameId")}/status`
       );
-      console.log("GameStatus", response.data);
-      if (response.data === "ANSWERING") {
+      console.log("GameStatus: ", response.data);
+      if(response.data==="WAITING" || response.data==="ANSWERING"){
         localStorage.setItem("myScore", 0);
         localStorage.setItem("roundNumber", 1);
         history.push(`/MultiGamePage/${gameId}/RoundCountPage/`);
       }
-      if (response.data === "DELETED" || response.data === "ENDED") {
-        backToLobby();
+      else if (response.data==="DELETED") {
+        localStorage.removeItem("gameId");
+        localStorage.removeItem("category");
+        localStorage.removeItem("totalRounds");
+        localStorage.removeItem("countdownTime");
+        localStorage.removeItem("playerNum");
+        localStorage.removeItem("isServer");
+        history.push(`/home`);
       }
-    } catch (error) {
+    }
+    catch (error) {
       toast.error(`Failed to fetch player in game(ID ${gameId})\n //change this
         ${error.response.data.message}`);
       console.log(handleError(error));
@@ -83,7 +86,7 @@ const CreatedGamePage = () => {
 
   useEffect(() => {
     if (isServer === "false") {
-      const interval1 = setInterval(fetchGameStatus, 3000);
+      const interval1 = setInterval(fetchGameStatus, 2000);
       return () => {
         clearInterval(interval1); // Clean up the interval on component unmount
       };
@@ -130,13 +133,18 @@ const CreatedGamePage = () => {
   }, []); */
 
   const leaveGame = async () => {
-    try{if (isServer === "true") {
-      const response = await api.delete(`/games/${gameId}`);
-      console.log("Delete game:", response.data);
-    } else {
-      const response = await api.delete(`/games/${gameId}/players/${userId}`);
-      console.log("Delete player:", response.data);
-    }}catch (error) {
+    try{
+      if (isServer === "true") {
+      await api.delete(`/games/${gameId}`);
+      await new Promise((resolve) => setTimeout(resolve, 500));
+      console.log(`Game ${gameId} deleted.`);
+      }
+      else {
+      await api.delete(`/games/${gameId}/players/${userId}`);
+      await new Promise((resolve) => setTimeout(resolve, 500));
+      console.log(`Player ${userId} deleted from Game ${gameId}`);
+      }
+    } catch (error) {
       toast.error(`Failed to fetch player in game(ID ${gameId})\n
         ${error.response.data.message}`);
       console.log(handleError(error));
